+++ src/USER-EFF/compute_temp_deform_eff.cpp
@@ -319,6 +319,6 @@ void ComputeTempDeformEff::restore_bias_all()
 
 double ComputeTempDeformEff::memory_usage()
 {
-  double bytes = maxbias * sizeof(double);
+  double bytes = (double)maxbias * sizeof(double);
   return bytes;
 }
