+++ src/USER-MISC/pair_list.cpp
@@ -409,10 +409,10 @@ double PairList::init_one(int, int)
 
 double PairList::memory_usage()
 {
-  double bytes = npairs * sizeof(int);
-  bytes += npairs * sizeof(list_parm_t);
+  double bytes = (double)npairs * sizeof(int);
+  bytes += (double)npairs * sizeof(list_parm_t);
   const int n = atom->ntypes+1;
-  bytes += n*(n*sizeof(int) + sizeof(int *));
-  bytes += n*(n*sizeof(double) + sizeof(double *));
+  bytes += (double)n*(n*sizeof(int) + sizeof(int *));
+  bytes += (double)n*(n*sizeof(double) + sizeof(double *));
   return bytes;
 }
