+++ src/VORONOI/compute_voronoi_atom.cpp
@@ -614,9 +614,9 @@ void ComputeVoronoi::processCell(voronoicell_neighbor &c, int i)
 
 double ComputeVoronoi::memory_usage()
 {
-  double bytes = size_peratom_cols * nmax * sizeof(double);
+  double bytes = (double)size_peratom_cols * nmax * sizeof(double);
   // estimate based on average coordination of 12
-  if (faces_flag) bytes += 12 * size_local_cols * nmax * sizeof(double);
+  if (faces_flag) bytes += (double)12 * size_local_cols * nmax * sizeof(double);
   return bytes;
 }
